Removed stub from Notification, replaced it with db

+++ Models/PrivateArea/Widgets/HeaderWidget.go
@@ -2,5 +2,5 @@ package Widgets
 
 type HeaderWidget struct {
 	Name          string `json:"name"`
-	Notifications int    `json:"notifications"`
+	Notifications int64  `json:"notifications"`
 }
